+++ src/abacus-backoffice/package.json
@@ -14,6 +14,7 @@
     "relay": "adeira-relay-compiler"
   },
   "dependencies": {
+    "@adeira/forms": "^0.1.0",
     "@adeira/hooks": "^0.3.0",
     "@adeira/icons": "^2.0.0",
     "@adeira/js": "^2.1.1",
