+++ src/forms/.babelrc.js
@@ -1,0 +1,5 @@
+// @flow strict
+
+module.exports = {
+  plugins: ['babel-plugin-fbt', 'babel-plugin-fbt-runtime'],
+};
