+++ src/forms/.eslintrc.js
@@ -1,0 +1,13 @@
+// @flow strict
+
+/* eslint-disable no-unused-vars */
+const OFF = 0;
+const WARN = 1;
+const ERROR = 2;
+/* eslint-enable no-unused-vars */
+
+module.exports = {
+  env: {
+    browser: true,
+  },
+};
