+++ src/forms/.npmignore
@@ -1,0 +1,2 @@
+__flowtests__
+__tests__
