+++ src/forms/jest-setup.js
@@ -1,0 +1,4 @@
+// @flow
+
+// eslint-disable-next-line import/no-extraneous-dependencies
+import '@testing-library/jest-dom';
