+++ src/forms/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@adeira/forms",
+  "description": "TKTK",
+  "homepage": "https://github.com/adeira/universe",
+  "bugs": "https://github.com/adeira/universe/issues",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:adeira/universe.git",
+    "directory": "src/forms"
+  },
+  "license": "MIT",
+  "version": "0.1.0",
+  "main": "./src/index.js",
+  "type": "commonjs",
+  "sideEffects": false,
+  "dependencies": {
+    "@adeira/js": "^2.1.1",
+    "@adeira/sx": "^0.29.1",
+    "@adeira/sx-design": "^0.29.0",
+    "@babel/runtime": "^7.18.3",
+    "fbt": "^1.0.0",
+    "recoil": "^0.7.4"
+  },
+  "devDependencies": {
+    "@fbtjs/default-collection-transform": "^1.0.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "babel-plugin-fbt": "^1.0.0",
+    "babel-plugin-fbt-runtime": "^1.0.0"
+  },
+  "peerDependencies": {
+    "@adeira/relay": "^6.0.0",
+    "react": "^18.2.0"
+  },
+  "scripts": {
+    "fbt:manifest": "fbt-manifest --src=src --src=pages --enum-manifest=translations/enum_manifest.json --src-manifest=translations/src_manifest.json",
+    "fbt:collect": "fbt-collect --options=__self --pretty --manifest < translations/src_manifest.json > translations/source_strings.json",
+    "fbt:translate": "fbt-translate --source-strings=translations/source_strings.json --pretty --translations translations/in/*.json --output-dir=translations/out --jenkins",
+    "fbt": "yarn run fbt:manifest && yarn run fbt:collect && yarn run fbt:translate"
+  }
+}
