+++ src/forms/src/index.js
@@ -1,0 +1,12 @@
+// @flow
+
+export { default as FormCheckboxList } from './FormCheckboxList';
+export { default as FormMultiSelect } from './FormMultiSelect';
+export { default as FormSelectOption } from './FormSelectOption';
+export { default as FormMultiUpload } from './FormMultiUpload';
+export { default as FormNumber } from './FormNumber';
+export { default as FormRoot } from './FormRoot';
+export { default as FormSelect } from './FormSelect';
+export { default as FormSubmit } from './FormSubmit';
+export { default as FormText } from './FormText';
+export { default as FormTextArea } from './FormTextArea';
