+++ src/forms/translations/enum_manifest.json
@@ -1,0 +1,1 @@
+{}
