+++ src/forms/translations/source_strings.json
@@ -1,0 +1,215 @@
+{
+ "phrases": [
+  {
+   "hashToLeaf": {
+    "DG4YwlsppzoioQ32fEtyQQ==": {
+     "text": "Something unexpected happened and server could not process the request! 🙈",
+     "desc": "generic failure message after creating a product"
+    }
+   },
+   "filepath": "src/FormSubmit.js",
+   "line_beg": 89,
+   "col_beg": 10,
+   "line_end": 92,
+   "col_end": 11,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "generic failure message after creating a product",
+     "text": "Something unexpected happened and server could not process the request! 🙈"
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "99TUdsKRfseIC6G04zBOKg==": {
+     "text": "This field is mandatory",
+     "desc": "mandatory field description"
+    }
+   },
+   "filepath": "src/private/BaseInputWrapper.js",
+   "line_beg": 33,
+   "col_beg": 21,
+   "line_end": 33,
+   "col_end": 90,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "mandatory field description",
+     "text": "This field is mandatory"
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "wE1bsb1gMGsz4iECjopB3w==": {
+     "text": "Field \"{form field label}\" is required. Please select one of the provided values.",
+     "desc": "value not selected validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 20,
+   "col_beg": 8,
+   "line_end": 23,
+   "col_end": 14,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "value not selected validation error message",
+     "text": "Field \"{form field label}\" is required. Please select one of the provided values."
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "4SRlvooaQB6It8KVa1bG9g==": {
+     "text": "Field \"{form field label}\" is required. Please fill in this field.",
+     "desc": "value missing validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 27,
+   "col_beg": 6,
+   "line_end": 30,
+   "col_end": 12,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "value missing validation error message",
+     "text": "Field \"{form field label}\" is required. Please fill in this field."
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "Wtx7FHIevcq36g3jquWXKg==": {
+     "text": "Field \"{form field label}\" must be negative.",
+     "desc": "range overflow validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 38,
+   "col_beg": 8,
+   "line_end": 41,
+   "col_end": 14,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "range overflow validation error message",
+     "text": "Field \"{form field label}\" must be negative."
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "S6jnfCvIypCDROzYeKb25A==": {
+     "text": "Field \"{form field label}\" has to have value smaller than {form field min value}.",
+     "desc": "range overflow validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 45,
+   "col_beg": 6,
+   "line_end": 48,
+   "col_end": 12,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "range overflow validation error message",
+     "text": "Field \"{form field label}\" has to have value smaller than {form field min value}."
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "x8iZhY9utb/UaEMs13hB8Q==": {
+     "text": "Field \"{form field label}\" must be positive.",
+     "desc": "range underflow validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 56,
+   "col_beg": 8,
+   "line_end": 59,
+   "col_end": 14,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "range underflow validation error message",
+     "text": "Field \"{form field label}\" must be positive."
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "qstwPQ9btvcdowx9na89+A==": {
+     "text": "Field \"{form field label}\" has to have value larger than {form field min value}.",
+     "desc": "range underflow validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 63,
+   "col_beg": 6,
+   "line_end": 66,
+   "col_end": 12,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "range underflow validation error message",
+     "text": "Field \"{form field label}\" has to have value larger than {form field min value}."
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "xBtaPdxqvTa9ZLg6/y0y9Q==": {
+     "text": "Please enter a valid value. The two nearest valid values are {low boundary} and {high boundary}.",
+     "desc": "step mismatch validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 79,
+   "col_beg": 6,
+   "line_end": 83,
+   "col_end": 12,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "step mismatch validation error message",
+     "text": "Please enter a valid value. The two nearest valid values are {low boundary} and {high boundary}."
+    },
+    "m": []
+   }
+  },
+  {
+   "hashToLeaf": {
+    "cXb0sWHtGL0NAVNk2AXyYw==": {
+     "text": "Field \"{form field label}\" is invalid.",
+     "desc": "generic validation error message"
+    }
+   },
+   "filepath": "src/private/getValidityStateMessage.js",
+   "line_beg": 89,
+   "col_beg": 4,
+   "line_end": 91,
+   "col_end": 10,
+   "project": "",
+   "jsfbt": {
+    "t": {
+     "desc": "generic validation error message",
+     "text": "Field \"{form field label}\" is invalid."
+    },
+    "m": []
+   }
+  }
+ ],
+ "childParentMappings": {}
+}
