+++ src/forms/translations/src_manifest.json
@@ -1,0 +1,1 @@
+{"translations/enum_manifest.json":["src/__tests__/FormCheckboxList.test.js","src/__tests__/FormMultiSelect.test.js","src/__tests__/FormMultiUpload.test.js","src/__tests__/FormNumber.test.js","src/__tests__/FormSelect.test.js","src/__tests__/FormText.test.js","src/__tests__/FormTextArea.test.js","src/FormSelect.js","src/FormSubmit.js","src/private/BaseInputWrapper.js","src/private/getValidityStateMessage.js"]}
